chore: move to express

+++ src/infra/http/dtos/books/create-book-body.ts
@@ -4,7 +4,7 @@ export const createBookBody = z.object({
   title: z.string(),
   author: z.string(),
   publishingCompany: z.string(),
-  publishingYear: z.number(),
-  numberOfPages: z.number(),
+  publishingYear: z.number().int(),
+  numberOfPages: z.number().int(),
   synopsis: z.string()
 })
